일단 여기까지... 로그인 회원가입 조건에 맞춰서 작동 다만 에러 메시지를 못띄움ㅠㅠㅠ 카카오톡 로그인 완료 대신에 댓글...미구현.. 테스트 코드 미구현

+++ src/main/java/com/example/projectblog/controller/BoardRestController.java
@@ -27,13 +27,14 @@ public class BoardRestController {
 
 
     @GetMapping("/api/board/{id}") //하나의 글 조회
-    public  BoardDto getOne(@PathVariable Integer id){
+    public  BoardDto getOne(@PathVariable Long id){
         return boardservice.getOne(id);
     }
 
 
     @PostMapping("/api/board") //글 작성
     public BoardDto create(@RequestBody BoardDto createBoard) {
+
         return boardservice.create(createBoard);
     }
